Model code refactor
