+++ models/contact.js
@@ -1,0 +1,58 @@
+const { Schema, model } = require("mongoose");
+const Joi = require("joi");
+
+const { handleMongooseError } = require("../helpers");
+
+const contactSchema = new Schema(
+  {
+    name: {
+      type: String,
+      required: [true, "Set name for contact"],
+    },
+    email: {
+      type: String,
+    },
+    phone: {
+      type: String,
+    },
+    favorite: {
+      type: Boolean,
+      default: false,
+    },
+  },
+  { versionKey: false }
+);
+
+contactSchema.post("save", handleMongooseError);
+
+const addSchema = Joi.object({
+  name: Joi.string().min(3).required().messages({
+    "any.required": "missing required name field",
+  }),
+  email: Joi.string()
+    .email({ minDomainSegments: 2, tlds: { allow: ["com", "net"] } })
+    .required()
+    .messages({
+      "any.required": "missing required email field",
+    }),
+  phone: Joi.string().min(6).required().messages({
+    "any.required": "missing required phone field",
+  }),
+  favorite: Joi.boolean(),
+});
+
+const updateFavoriteSchema = Joi.object({
+  favorite: Joi.boolean().required(),
+});
+
+const schemas = {
+  addSchema,
+  updateFavoriteSchema,
+};
+
+const Contact = model("contact", contactSchema);
+
+module.exports = {
+  Contact,
+  schemas,
+};
