--- models/contacts.js
@@ -1,60 +1,0 @@
-const fs = require("fs/promises");
-const path = require("path");
-const { nanoid } = require("nanoid");
-
-const contactsPath = path.resolve(__dirname, "contacts.json");
-
-const listContacts = async () => {
-  const data = await fs.readFile(contactsPath, { encoding: "utf8" });
-  const contactsList = JSON.parse(data);
-  return contactsList;
-};
-
-const getContactById = async (contactId) => {
-  const contactsList = await listContacts();
-  const contactById = contactsList.find((contact) => contact.id === contactId);
-  return contactById || null;
-};
-
-const removeContact = async (contactId) => {
-  const contactsList = await listContacts();
-  const index = contactsList.findIndex((item) => item.id === contactId);
-  if (index === -1) {
-    return null;
-  }
-  const [delContact] = contactsList.splice(index, 1);
-  await fs.writeFile(contactsPath, JSON.stringify(contactsList, null, 2));
-  return delContact;
-};
-
-const addContact = async (body) => {
-  const newContact = {
-    id: nanoid(),
-    ...body,
-  };
-  const contactsList = await listContacts();
-
-  contactsList.push(newContact);
-
-  await fs.writeFile(contactsPath, JSON.stringify(contactsList, null, 2));
-  return newContact;
-};
-
-const updateContact = async (id, body) => {
-  const contactsList = await listContacts();
-  const index = contactsList.findIndex((item) => item.id === id);
-  if (index === -1) {
-    return null;
-  }
-  contactsList[index] = { id, ...body };
-  await fs.writeFile(contactsPath, JSON.stringify(contactsList, null, 2));
-  return contactsList[index];
-};
-
-module.exports = {
-  listContacts,
-  getContactById,
-  removeContact,
-  addContact,
-  updateContact,
-};
